Enabled 1-day scans
Allowed -eday to not be set in order to allow for one day scans.
Previously you had to set -sday and -eday with -eday being one day
later.

However, xrange() is not inclusive of the second value given which
leads to confusion over setting the end day and not having the end day
returned in the results.